Added documentation to module, download the google mobility and population data now if they don't exist. Added data_processing_config.yaml and metadata.yml

+++ extractors/aggregate_flows_from_google_data.py
@@ -1,3 +1,36 @@
+"""
+This script will create the movement multipliers using:
+- A populations estimate for each local authority in Scotland from 2018 (mid year) https://www2.gov.scot/Resource/0046/00462936.csv
+- Google mobility data giving the number of movements between geographic regions in 2020 https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv
+- A table mapping the iso_3166_2 name of each local authority to the local authority name and code
+   compiled by hand, Jess Enright, 30 June 2020
+
+Python requirements:
+- pathlib
+- pandas
+- zipfile
+- ftplib.FTP  ??? Only if the hand generated lookup table is hosted on the FTP server.
+- urllib.request
+- data_pipeline_api
+
+How to run this module:
+This script uses the data_pipeline_api.data_processing_api, and therefore makes use of both
+- data_processing_config.yaml, and
+- metadata.yaml
+metadata.yaml should include something like:
+```
+- doi_or_unique_name: mid-year-pop-est-18-tabs_Table 2.csv
+  filename: human/external/mid-year-pop-est-18-tabs_Table 2.csv
+- doi_or_unique_name: Global_Mobility_Report.csv
+  filename: human/external/Global_Mobility_Report.csv
+```
+This script can be run with
+```
+python aggregate_flows_from_google_data.py
+```
+The script generates .h5 files in generated_sns_products/movement_multiplier
+"""
+
 from pathlib import Path
 import pandas as pd
 import urllib.request
@@ -5,27 +38,38 @@ from data_pipeline_api.data_processing_api import DataProcessingAPI
 from ftplib import FTP
 import zipfile
 
+
 config_filename =  Path(__file__).parent / "data_processing_config.yaml"
 uri = "data_processing_uri"
 git_sha = "data_processing_git_sha"
+data_path = Path(__file__).parent / "human/external"
 
 
 def download_pop_table():
     """
-    Download the population data from an external source, using only the Area1 data and the first 3 columns
+    Download the population data from an external source if it doesn't exists in data_path, using only the Area1 data and the first 3 columns
     the LA code, name and population, removing the commas in the population numbers.
     :return: A dataframe containing the local authority code, name and population
              and the total polulation
     """
-    url = "https://www.nrscotland.gov.uk/files//statistics/population-estimates/mid-18/mid-year-pop-est-18-tabs.zip"
-    zip_filename = "mid-year-pop-est-18-tabs.zip"
-    urllib.request.urlretrieve(
-        url, zip_filename
-    )
 
     population_table = "mid-year-pop-est-18-tabs_Table 2.csv"
-    with zipfile.ZipFile(zip_filename, 'r') as zip_ref:
-        zip_ref.extractall(".", members=[population_table])
+
+
+    # If the population table doesn't exist download it.
+    if not Path(data_path / population_table).exists():
+        print(f"Could not find {data_path}/{population_table}, downloading it")
+        url = "https://www.nrscotland.gov.uk/files//statistics/population-estimates/mid-18/mid-year-pop-est-18-tabs.zip"
+        zip_filename = "mid-year-pop-est-18-tabs.zip"
+        urllib.request.urlretrieve(
+            url, zip_filename
+        )
+
+        with zipfile.ZipFile(zip_filename, 'r') as zip_ref:
+            zip_ref.extractall(data_path, members=[population_table])
+
+        # clean up (i.e. remove) the downloaded datafile(s)
+        Path(zip_filename).unlink(missing_ok=False)
 
     with DataProcessingAPI(config_filename, uri=uri, git_sha=git_sha) as api:
         with api.read_external_object(population_table) as file:
@@ -34,10 +78,6 @@ def download_pop_table():
             dfPop['population'] = dfPop['population'].str.replace(',', '').astype(int)
 
     total_population =  dfPop['population'].sum()
-
-    # clean up (i.e. remove) the downloaded datafile(s)
-    Path(zip_filename).unlink(missing_ok=False)
-    Path(population_table).unlink(missing_ok=False)
 
     return dfPop, total_population
 
@@ -99,7 +139,6 @@ def download_lookup_table():
     """
 
     dfLookup = pd.read_csv(lookup_table, low_memory=False)
-    print(dfLookup)
     dfLookup['full_iso_code'] = "GB-" + dfLookup.iso_3166_2
     codes_of_interest = list(dfLookup['full_iso_code'])
     dfLookup = dfLookup[['full_iso_code', 'la_code']]
@@ -111,23 +150,25 @@ def download_lookup_table():
 
 def download_google_mogility_data(la_list):
     """
-    Downloads the mobility data from google for Scottish local authorities
-    :param: The list of local åuthorities ﬅor which we want the movement data
+    Downloads the mobility data from google for Scottish local authorities if it doesn't exists in data_path
+    :param: The list of local åuthorities for which we want the movement data
     :return: A dataframe containing the full ISO code (GB-iso_3166_2) and the corresponding local authority code
     """
-    url = "https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv"
+
     google_mobility_table = "Global_Mobility_Report.csv"
-    urllib.request.urlretrieve(
-            url, google_mobility_table
-            )
+
+    # If the population table doesn't exist download it.
+    if not Path(data_path / google_mobility_table).exists():
+        print(f"Could not find {data_path}/{google_mobility_table}, downloading it")
+        url = "https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv"
+        urllib.request.urlretrieve(
+                url, data_path / google_mobility_table
+                )
 
     with DataProcessingAPI(config_filename, uri=uri, git_sha=git_sha) as api:
-        with api.read_external_object(population_table) as file:
+        with api.read_external_object(google_mobility_table) as file:
             justScotGoogle = pd.read_csv(google_mobility_table, low_memory=False)
             justScotGoogle = justScotGoogle[justScotGoogle['iso_3166_2_code'].isin(la_list) == True]
-
-    # clean up (i.e. remove) the downloaded datafile(s)
-    Path(google_mobility_table).unlink(missing_ok=False)
 
     return justScotGoogle
 
@@ -166,6 +207,12 @@ def main():
 
     # TODO - upload this to the database; human/movement-multipliers/1/data.h5
     dfScotGoogle.to_csv('movement_multiplier.csv')
-    
+    with DataProcessingAPI(config_filename, uri=uri, git_sha=git_sha) as api:
+        api.write_table(
+            "generated_sns_products/movement_multiplier",
+            "movement_multiplier",
+            dfScotGoogle,
+            )    
+
 if __name__ == "__main__":
     main()
